log:	709533 Problem: When creating a new Templates on a Works Order, PEM Record etc. which already has a template or document associated to it, the 'Create from Template' option is greyed out unless the highlighted field is changed to a new row under the existing ones. This was not neccesary in V2 and has led to some User's not associating documents or creating templates (especially as the latest documents/templates are shown at the top of the list in date order).
Solution:
button was enabled and disabled at different points in the form, I've removed all this code from program unit set_field_attribs so that the button is always enabled. The when-button-pressed trigger behind the button DAS.DASKEYTEMPLATE checks to see if the user is adding a new record. If they are not it forces a new record and then calls the form to allow the user to select the template they want to use (module doc0200).

git-svn-id: svn://127.0.0.1/Core@156 17c802ae-cbca-534c-9e8e-6d19a4926c45