Log 711042
Problem: Unable to resolve reference to item RES_CONTROL.LOV_OK error message displayed when the checkbox is ticked/unticked.

Fix: The above item did not exist in the RES_CONTROL block but there were several references to it via a set_item_enabled function. These references were added as a result of work done to NM0572 v1.28 to enable/disable the 'ok' button on that form depending on the value of a checkbox. Since the 'ok' button doesn't exist in NM0573 it was decided to comment out the references to the LOV_OK item in the RES_CONTROL block.

git-svn-id: svn://127.0.0.1/Core@717 17c802ae-cbca-534c-9e8e-6d19a4926c45