Log 715234
Problem: Error message is displayed when entering an asset and an admin unit via the LOV. Problem caused when the admin unit is validate. It isn't being restricted by the admin type for the specific asset type 

Fix: Modified the when-validate-item trigger on the nau_code (qry_results block) to restrict validation by the admin type.

git-svn-id: svn://127.0.0.1/Core@1302 17c802ae-cbca-534c-9e8e-6d19a4926c45