log 711438
problem: Problems displaying selected assets that have been identified in the search facility in locator. The search lists the assets on the LHS and you are able to tick the appropriate asset but when you click on the globe to display selected nothing happens. 

solution: created procedure send_multiple and set_many_theme in program unit map_inventory to cater for multiple selected assets.  

git-svn-id: svn://127.0.0.1/Core@1442 17c802ae-cbca-534c-9e8e-6d19a4926c45