�714254 � its data that is configured locally at Kentucky. It is a report where they have a date parameter � the LOV will not order by the date � only by the character representation of the date.�
This one requires a change to GRI0200 which I�ll get into 4052.

When the extended lov is called it constructs the query to populate the form with like this�

SELECT
  SHOWN,
  DESCR,
  VALUE
FROM
  (SELECT to_char(NFI_DATE_LOADED) shown,
to_char(NFI_DATE_LOADED) descr,
to_char(NFI_DATE_LOADED) value 
 FROM NSG_LOAD_FILES
 WHERE 1=1 order by NFI_DATE_LOADED DESC)
WHERE
  (UPPER(SHOWN) LIKE '%'
   OR
   UPPER(DESCR) LIKE '%')
ORDER BY
  1


I found that when GRI0200 calls the extended lov it always passes over a parameter value that tells it to order the outer select on the first 
column, which is always a varchar.  Although the inner select uses the GRI_PARAMS.GP_ORDER value when it�s constructed, the to_char buggers it up.

When you invoke the extended LOV you can actually tell it not to apply an order by to the outer select that it constructs.  

So, if GRI_PARAMS.GP_ORDER is null I�ll get it to order on the first column, else tell it not to order the outer select.

git-svn-id: svn://127.0.0.1/Core@1673 17c802ae-cbca-534c-9e8e-6d19a4926c45