Log 717827 Problem - The 'surveyed by' field was not being populated. changed :qry_results post-query trigger to populate this field.
SC 01/03/2011 - Added PVCS tag nm_4200_fix1 to bring it inline with what down load manager reports went out in 4200 fix1.  However i can't find fix 1 on the download site and the log assocaited with this fix i can't find in problem manager.  Furthermore the next revision 2.4.1.1 is also label as log 717827, which makes me think that this version (2.4.1.1) is actually the version for the fix not this one (2.4.1.0)

git-svn-id: svn://127.0.0.1/Core@2282 17c802ae-cbca-534c-9e8e-6d19a4926c45